fix: consolidate factorydefinition tests under package file cap
Merge activation and coverage regression tests into read_test.go,
save_test.go, and persist_test.go so pkg/factorydefinition/service
stays within the 15-file limit while preserving 80%+ coverage.
